Simple object in scene
Managed to get building to my Android phone working.
New: Made a Prefab for the Chest with a simple script intended to open and close the chest. But decided to laern abit more of Unity. It would make sese to have an interface for things that can open, close and potentially be locked.